server/requirements.txt
### Librerias actualizadas con Flask 2.2.2 y Python 3.11.0 ###
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==0.21.1
six==1.16.0
typish==1.9.3
Werkzeug==2.2.2